+++ library/MultiRegions/DisContField3D.cpp
@@ -675,6 +675,7 @@ using namespace std;
             // Finally the three sets locVerts, locEdges and locFaces store any
             // vertices, edges and faces that belong to a periodic composite and
             // lie on this process.
+            map<int,RotPeriodicInfo>                       rotComp;
             map<int,int>                                   perComps;
             map<int,vector<int> >                          allVerts;
             map<int,SpatialDomains::PointGeomVector>       allCoord;
@@ -708,6 +709,7 @@ using namespace std;
             // it, then fill out the all* maps.
             for (auto &it : bregions)
             {
+                
                 locBCond = GetBoundaryCondition(
                     bconditions, it.first, variable);
 
@@ -745,8 +747,61 @@ using namespace std;
                     cId2 = bndRegOrder.find(region2ID)->second[0];
                 }
 
+                // check to see if boundary is rotationally aligned
+                if(boost::icontains(locBCond->GetUserDefined(),"Rotated"))
+                {
+                    vector<string> tmpstr;
+                    
+                    boost::split(tmpstr,locBCond->GetUserDefined(),
+                                 boost::is_any_of(":"));
+                    
+                    if(boost::iequals(tmpstr[0],"Rotated"))
+                    {
+                        ASSERTL1(tmpstr.size() > 2,
+                                 "Expected Rotated user defined string to "
+                                 "contain direction and rotation angle "
+                                 "and optionally a tolerance, "
+                                 "i.e. Rotated:dir:PI/2:1e-6");
+
+
+                        ASSERTL1((tmpstr[1] == "x")||(tmpstr[1] == "y")
+                                  ||(tmpstr[1] == "z"), "Rotated Dir is "
+                                  "not specified as x,y or z");
+                        
+                        RotPeriodicInfo RotInfo;
+                        RotInfo.m_dir = (tmpstr[1] == "x")? 0:
+                            (tmpstr[1] == "y")? 1:2;
+
+                        LibUtilities::AnalyticExpressionEvaluator strEval;
+                        int ExprId = strEval.DefineFunction("", tmpstr[2]);
+                        RotInfo.m_angle = strEval.Evaluate(ExprId);
+
+                        if(tmpstr.size() == 4)
+                        {
+                            try {
+                                RotInfo.m_tol = boost::lexical_cast
+                                    <NekDouble>(tmpstr[3]);
+                            }
+                            catch (...) {
+                                NEKERROR(ErrorUtil::efatal,
+                                         "failed to cast tolerance input "
+                                         "to a double value in Rotated"
+                                         "boundary information");
+                             }
+                        }
+                        else
+                        {
+                            RotInfo.m_tol = 1e-8;
+                        }
+                        rotComp[cId1] = RotInfo;
+                    }
+                }
+
                 SpatialDomains::CompositeSharedPtr c = it.second->begin()->second;
+
                 vector<unsigned int> tmpOrder;
+
+                // store the rotation info of this 
                 
                 // From the composite, we now construct the allVerts, allEdges
                 // and allCoord map so that they can be transferred across
@@ -833,6 +888,50 @@ using namespace std;
             Array<OneD, int> vertcounts(n,0);
             Array<OneD, int> faceoffset(n,0);
             Array<OneD, int> vertoffset(n,0);
+
+            Array<OneD, int> rotcounts(n,0);
+            Array<OneD, int> rotoffset(n,0);
+
+            rotcounts[p] = rotComp.size();
+            vComm->AllReduce(rotcounts, LibUtilities::ReduceSum);
+            int totrot  = Vmath::Vsum(n,rotcounts,1);
+
+            if(totrot)
+            {
+                for (i = 1; i < n ; ++i)
+                {
+                    rotoffset[i] = rotoffset[i-1] + rotcounts[i-1];
+                }
+
+                Array<OneD, int> compid(totrot,0);
+                Array<OneD, int> rotdir(totrot,0);
+                Array<OneD, NekDouble> rotangle(totrot,0.0);
+                Array<OneD, NekDouble> rottol(totrot,0.0);
+
+                // fill in rotational informaiton
+                auto rIt = rotComp.begin();
+                
+                for(i = 0; rIt != rotComp.end(); ++rIt)
+                {
+                    compid  [rotoffset[p] + i  ] = rIt->first; 
+                    rotdir  [rotoffset[p] + i  ] = rIt->second.m_dir; 
+                    rotangle[rotoffset[p] + i  ] = rIt->second.m_angle; 
+                    rottol  [rotoffset[p] + i++] = rIt->second.m_tol; 
+                }
+
+                vComm->AllReduce(compid, LibUtilities::ReduceSum);
+                vComm->AllReduce(rotdir, LibUtilities::ReduceSum);
+                vComm->AllReduce(rotangle, LibUtilities::ReduceSum);
+                vComm->AllReduce(rottol, LibUtilities::ReduceSum);
+
+                // Fill in full rotational composite list
+                for(i =0; i < totrot; ++i)
+                {
+                    RotPeriodicInfo rinfo(rotdir[i],rotangle[i], rottol[i]);
+
+                    rotComp[compid[i]] = rinfo; 
+                }
+            }
 
             // First exchange the number of faces on each process.
             facecounts[p] = locFaces.size();
@@ -1063,6 +1162,9 @@ using namespace std;
 
             map<int,int> allCompPairs;
             
+            // Collect composite ides of each periodic face for use if rotation is required
+            map<int,int> fIdToCompId;
+
             // Finally we have enough information to populate the periodic
             // vertex, edge and face maps. Begin by looping over all pairs of
             // periodic composites to determine pairs of periodic faces.
@@ -1092,6 +1194,7 @@ using namespace std;
                 // process.
                 map<int,int> compPairs;
 
+
                 ASSERTL0(compOrder.count(id1) > 0,
                          "Unable to find composite "+id1s+" in order map.");
                 ASSERTL0(compOrder.count(id2) > 0,
@@ -1118,10 +1221,14 @@ using namespace std;
                         ASSERTL0(compPairs[eId2] == eId1, "Pairing incorrect");
                     }
                     compPairs[eId1] = eId2;
+                    
+                    // store  a map of face ids to composite ids
+                    fIdToCompId[eId1] = id1;
+                    fIdToCompId[eId2] = id2;
                 }
 
                 // Now that we have all pairs of periodic faces, loop over the
-                // ones local to this process and populate face/edge/vertex
+                // ones local on this process and populate face/edge/vertex
                 // maps.
                 for (auto &pIt : compPairs)
                 {
@@ -1151,6 +1258,19 @@ using namespace std;
                     // different going from face1->face2 instead of face2->face1
                     // (check this).
                     StdRegions::Orientation o;
+                    bool rotbnd = false;
+                    int  dir;
+                    NekDouble angle,sign;
+                    NekDouble tol = 1e-8;
+
+                    // check to see if perioid boundary is rotated
+                    if(rotComp.count(fIdToCompId[pIt.first]))
+                    {
+                        rotbnd = true;
+                        dir   = rotComp[fIdToCompId[pIt.first]].m_dir;
+                        angle = rotComp[fIdToCompId[pIt.first]].m_angle;
+                        tol   = rotComp[fIdToCompId[pIt.first]].m_tol;
+                    }
 
                     // Record periodic faces.
                     for (i = 0; i < 2; ++i)
@@ -1163,16 +1283,21 @@ using namespace std;
                         // Reference to the other face.
                         int other = (i+1) % 2;
 
+                        // angle is set up for i = 0 to i = 1
+                        sign = (i == 0)? 1.0:-1.0;
+                        
                         // Calculate relative face orientation.
                         if (tmpVec[0].size() == 3)
                         {
                             o = SpatialDomains::TriGeom::GetFaceOrientation(
-                                tmpVec[i], tmpVec[other]);
+                                   tmpVec[i], tmpVec[other],
+                                   rotbnd, dir, sign*angle, tol);
                         }
                         else
                         {
                             o = SpatialDomains::QuadGeom::GetFaceOrientation(
-                                tmpVec[i], tmpVec[other]);
+                                   tmpVec[i], tmpVec[other],
+                                   rotbnd,dir,sign*angle,tol);
                         }
 
                         // Record face ID, orientation and whether other face is
@@ -1189,16 +1314,21 @@ using namespace std;
                     {
                         int other = (i+1) % 2;
 
+                        // angle is set up for i = 0 to i = 1
+                        sign = (i == 0)? 1.0:-1.0;
+
                         // Calculate relative face orientation.
                         if (tmpVec[0].size() == 3)
                         {
                             o = SpatialDomains::TriGeom::GetFaceOrientation(
-                                tmpVec[i], tmpVec[other]);
+                                   tmpVec[i], tmpVec[other], rotbnd, dir,
+                                   sign*angle, tol);
                         }
                         else
                         {
                             o = SpatialDomains::QuadGeom::GetFaceOrientation(
-                                tmpVec[i], tmpVec[other]);
+                                   tmpVec[i], tmpVec[other], rotbnd, dir,
+                                   sign*angle, tol);
                         }
 
                         if (nFaceVerts == 3)
@@ -1271,15 +1401,20 @@ using namespace std;
                     {
                         int other = (i+1) % 2;
 
+                        // angle is set up for i = 0 to i = 1
+                        sign = (i == 0)? 1.0:-1.0;
+
                         if (tmpVec[0].size() == 3)
                         {
                             o = SpatialDomains::TriGeom::GetFaceOrientation(
-                                tmpVec[i], tmpVec[other]);
+                                   tmpVec[i], tmpVec[other], rotbnd, dir,
+                                   sign*angle, tol);
                         }
                         else
                         {
                             o = SpatialDomains::QuadGeom::GetFaceOrientation(
-                                tmpVec[i], tmpVec[other]);
+                                   tmpVec[i], tmpVec[other], rotbnd, dir,
+                                   sign*angle, tol);
                         }
 
                         vector<int> per1 = edgeMap[ids[i]];
@@ -1342,6 +1477,11 @@ using namespace std;
                 pairOffsets[i] = pairOffsets[i-1] + pairSizes[i-1];
             }
 
+
+            ASSERTL1(allCompPairs.size() == fIdToCompId.size(),
+                     "At this point the size of allCompPairs "
+                     "should have been the same as fIdToCompId");
+
             Array<OneD, int> first (totPairSizes, 0);
             Array<OneD, int> second(totPairSizes, 0);
 
@@ -1352,7 +1492,7 @@ using namespace std;
                 first [cnt  ] = pIt.first;
                 second[cnt++] = pIt.second;
             }
-
+            
             vComm->AllReduce(first,  LibUtilities::ReduceSum);
             vComm->AllReduce(second, LibUtilities::ReduceSum);
 
@@ -1363,18 +1503,64 @@ using namespace std;
                 allCompPairs[first[cnt]] = second[cnt];
             }
 
+            // make global list of faces to composite ids if rotComp is non-zero
+            
+            if(rotComp.size())
+            {
+                Vmath::Zero(totPairSizes,first,1);
+                Vmath::Zero(totPairSizes,second,1);
+                
+                cnt = pairOffsets[p];
+                
+                for (auto &pIt : fIdToCompId)
+                {
+                    first [cnt  ] = pIt.first;
+                    second[cnt++] = pIt.second;
+                }
+                
+                vComm->AllReduce(first,  LibUtilities::ReduceSum);
+                vComm->AllReduce(second, LibUtilities::ReduceSum);
+                
+                fIdToCompId.clear();
+                
+                for(cnt = 0; cnt < totPairSizes; ++cnt)
+                {
+                    fIdToCompId[first[cnt]] = second[cnt];
+                }
+            }
+
+            // also will need an edge id to composite id at end of routine
+            map<int,int> eIdToCompId; 
+
             // Search for periodic vertices and edges which are not
             // in a periodic composite but lie in this process. First,
             // loop over all information we have from other
             // processors.
             for (cnt = i = 0; i < totFaces; ++i)
             {
+                bool rotbnd = false;
+                int dir;
+                NekDouble angle;
+                NekDouble tol = 1e-8;
+                
                 int faceId    = faceIds[i];
-
+                
                 ASSERTL0(allCompPairs.count(faceId) > 0,
                          "Unable to find matching periodic face.");
 
                 int perFaceId = allCompPairs[faceId];
+
+                // check to see if periodic boundary is rotated
+                ASSERTL1(fIdToCompId.count(faceId) > 0,"Face " +
+                         boost::lexical_cast<string>(faceId) +
+                         " not found in fIdtoCompId map");
+                if(rotComp.count(fIdToCompId[faceId]))
+                {
+                    rotbnd = true;
+                    dir   = rotComp[fIdToCompId[faceId]].m_dir;
+                    angle = rotComp[fIdToCompId[faceId]].m_angle;
+                    tol   = rotComp[fIdToCompId[faceId]].m_tol;
+                }
 
                 for (j = 0; j < faceVerts[i]; ++j, ++cnt)
                 {
@@ -1384,21 +1570,22 @@ using namespace std;
 
                     if (perId == periodicVerts.end())
                     {
-
-                        // This vertex is not included in the map. Figure out which
-                        // vertex it is supposed to be periodic with. perFaceId is
-                        // the face ID which is periodic with faceId. The logic is
-                        // much the same as the loop above.
+                        
+                        // This vertex is not included in the
+                        // map. Figure out which vertex it is supposed
+                        // to be periodic with. perFaceId is the face
+                        // ID which is periodic with faceId. The logic
+                        // is much the same as the loop above.
                         SpatialDomains::PointGeomVector tmpVec[2]
                             = { coordMap[faceId], coordMap[perFaceId] };
                         
                         int nFaceVerts = tmpVec[0].size();
                         StdRegions::Orientation o = nFaceVerts == 3 ? 
                             SpatialDomains::TriGeom::GetFaceOrientation(
-                                                                        tmpVec[0], tmpVec[1]) :
+                                  tmpVec[0], tmpVec[1], rotbnd, dir, angle, tol):
                             SpatialDomains::QuadGeom::GetFaceOrientation(
-                                                                         tmpVec[0], tmpVec[1]);
-                        
+                                   tmpVec[0], tmpVec[1], rotbnd, dir, angle, tol);
+                            
                         // Use vmap to determine which vertex of the other face
                         // should be periodic with this one.
                         int perVertexId = vertMap[perFaceId][vmap[nFaceVerts][o][j]];
@@ -1414,6 +1601,12 @@ using namespace std;
                     int eId = edgeIds[cnt];
 
                     perId = periodicEdges.find(eId);
+
+                    // this map is required at very end to determine rotation of edges. 
+                    if(rotbnd)
+                    {
+                        eIdToCompId[eId] = fIdToCompId[faceId];
+                    }
                     
                     if (perId == periodicEdges.end())
                     {
@@ -1428,20 +1621,27 @@ using namespace std;
                         int nFaceEdges = tmpVec[0].size();
                         StdRegions::Orientation o = nFaceEdges == 3 ? 
                             SpatialDomains::TriGeom::GetFaceOrientation(
-                            tmpVec[0], tmpVec[1]) :
+                                        tmpVec[0], tmpVec[1], rotbnd, dir, angle, tol):
                         SpatialDomains::QuadGeom::GetFaceOrientation(
-                            tmpVec[0], tmpVec[1]);
+                                        tmpVec[0], tmpVec[1], rotbnd, dir, angle, tol);
                         
                         // Use emap to determine which edge of the other
                         // face should be periodic with this one.
                         int perEdgeId = edgeMap[perFaceId][emap[nFaceEdges][o][j]];
-                        
                         
                         PeriodicEntity ent(perEdgeId,
                                            StdRegions::eForwards,
                                            locEdges.count(perEdgeId) > 0);
                         
                         periodicEdges[eId].push_back(ent);
+                        
+
+                        // this map is required at very end to
+                        // determine rotation of edges.
+                        if(rotbnd)
+                        {
+                            eIdToCompId[perEdgeId] = fIdToCompId[perFaceId];
+                        }
                     }
                 }
             }
@@ -1518,12 +1718,27 @@ using namespace std;
             // Loop over periodic edges to determine relative edge orientations.
             for (auto &perIt : periodicEdges)
             {
+                bool rotbnd = false;
+                int dir;
+                NekDouble angle;
+                NekDouble tol = 1e-8;
+                
+
                 // Find edge coordinates
                 auto eIt = eIdMap.find(perIt.first);
                 SpatialDomains::PointGeom v[2] = {
                     *vCoMap[eIt->second.first],
                     *vCoMap[eIt->second.second]
                 };
+
+                // check to see if perioid boundary is rotated
+                if(rotComp.count(eIdToCompId[perIt.first]))
+                {
+                    rotbnd = true;
+                    dir   = rotComp[eIdToCompId[perIt.first]].m_dir;
+                    angle = rotComp[eIdToCompId[perIt.first]].m_angle;
+                    tol   = rotComp[eIdToCompId[perIt.first]].m_tol;
+                }
 
                 // Loop over each edge, and construct a vector that takes us
                 // from one vertex to another. Use this to figure out which
@@ -1537,39 +1752,66 @@ using namespace std;
                         *vCoMap[eIt->second.second]
                     };
 
-                    NekDouble cx = 0.5*(w[0](0)-v[0](0)+w[1](0)-v[1](0));
-                    NekDouble cy = 0.5*(w[0](1)-v[0](1)+w[1](1)-v[1](1));
-                    NekDouble cz = 0.5*(w[0](2)-v[0](2)+w[1](2)-v[1](2));
-
                     int vMap[2] = {-1,-1};
-                    for (j = 0; j < 2; ++j)
+                    if(rotbnd)
                     {
-                        NekDouble x = v[j](0);
-                        NekDouble y = v[j](1);
-                        NekDouble z = v[j](2);
-                        for (k = 0; k < 2; ++k)
-                        {
-                            NekDouble x1 = w[k](0)-cx;
-                            NekDouble y1 = w[k](1)-cy;
-                            NekDouble z1 = w[k](2)-cz;
+                        
+                        SpatialDomains::PointGeom r;
 
-                            if (sqrt((x1-x)*(x1-x)+(y1-y)*(y1-y)+(z1-z)*(z1-z))
-                                    < 1e-8)
+                        r.Rotate(v[0],dir,angle);
+                        
+                        if(r.dist(w[0])< tol)
+                        {
+                            vMap[0] = 0;
+                        }
+                        else
+                        {
+                            r.Rotate(v[1],dir,angle);
+                            if(r.dist(w[0]) < tol)
                             {
-                                vMap[k] = j;
-                                break;
+                                vMap[0] = 1;
+                            }
+                            else
+                            {
+                                ASSERTL0(false,"Unable to align rotationally periodic edge vertex");
                             }
                         }
                     }
+                    else // translation test
+                    {
+                        NekDouble cx = 0.5*(w[0](0)-v[0](0)+w[1](0)-v[1](0));
+                        NekDouble cy = 0.5*(w[0](1)-v[0](1)+w[1](1)-v[1](1));
+                        NekDouble cz = 0.5*(w[0](2)-v[0](2)+w[1](2)-v[1](2));
 
-                    // Sanity check the map.
-                    ASSERTL0(vMap[0] >= 0 && vMap[1] >= 0,
-                             "Unable to align periodic vertices.");
-                    ASSERTL0((vMap[0] == 0 || vMap[0] == 1) &&
-                             (vMap[1] == 0 || vMap[1] == 1) &&
-                             (vMap[0] != vMap[1]),
-                             "Unable to align periodic vertices.");
-
+                        for (j = 0; j < 2; ++j)
+                        {
+                            NekDouble x = v[j](0);
+                            NekDouble y = v[j](1);
+                            NekDouble z = v[j](2);
+                            for (k = 0; k < 2; ++k)
+                            {
+                                NekDouble x1 = w[k](0)-cx;
+                                NekDouble y1 = w[k](1)-cy;
+                                NekDouble z1 = w[k](2)-cz;
+                                
+                                if (sqrt((x1-x)*(x1-x)+(y1-y)*(y1-y)+(z1-z)*(z1-z))
+                                    < 1e-8)
+                                {
+                                    vMap[k] = j;
+                                    break;
+                                }
+                            }
+                        }
+                    
+                        // Sanity check the map.
+                        ASSERTL0(vMap[0] >= 0 && vMap[1] >= 0,
+                                 "Unable to align periodic edge vertex.");
+                        ASSERTL0((vMap[0] == 0 || vMap[0] == 1) &&
+                                 (vMap[1] == 0 || vMap[1] == 1) &&
+                                 (vMap[0] != vMap[1]),
+                                 "Unable to align periodic edge vertex.");
+                    }
+                    
                     // If 0 -> 0 then edges are aligned already; otherwise
                     // reverse the orientation.
                     if (vMap[0] != 0)