+++ utilities/NekMesh/InputModules/InputNekpp.cpp
@@ -268,6 +268,7 @@ void InputNekpp::Process()
             {
                 EdgeSharedPtr edg = eIdMap[geomIt->GetGlobalID()];
                 E->SetVolumeNodes(edg->m_edgeNodes);
+                E->SetCurveType(edg->m_curveType);
             }
 
             if (dim > 1)
@@ -285,6 +286,7 @@ void InputNekpp::Process()
                 {
                     FaceSharedPtr fac = fIdMap[geomIt->GetGlobalID()];
                     E->SetVolumeNodes(fac->m_faceNodes);
+                    E->SetCurveType(fac->m_curveType);
                 }
             }